requirements.txt
streamlit
pytest
pydeck
pandas
python-dotenv